+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Waypoint delete UX made a little easier. Explicit button in the waypoint activity, rather than a long-press
 - Notification permissions are requested sensibly on Android 13+
 - Google Play Store build numbers will now just increment from 40800000, rather than reflect the version string
+- Background location permission! (Hurray!)
 
 ### Bug fixes